Pledge Meta: Update post title when `org-name` gets updated
The value in the `org-name` field is also used as the title of the
pledge post. If this value changes (via the Manage Pledge form, for
example), the post title should be updated as well.

There won't be any public way to change the post title directly, so
we shouldn't need to implement an auto-update of the post meta value.