Remove unnecessary meta config and prefixes
The meta config introduced in 2927532 are unnecessary because the
contributors associated with a pledge are not stored as post meta data,
but rather as posts of a separate post type. The added `org-` prefixes
are also unnecessary because there is already a `5ftf` prefix for all
the fields. The `org-` is used for some fields where the data is a
property of the organization making the pledge (org-name, org-email).
Contributors are not really a property of the organization.