Log: Record when contributors are added to a pledge
This also updates the function that creates new contributor posts. It
replaces `create_new_contributor` with `add_pledge_contributors`. This
way we can capture a batch of contributor additions all in one event by
collecting all the `wp_insert_post` results in an array and including
it as an action parameter.