Stats: Clarify confusing "sponsored" stats
This removes the `_total_sponsored_hours` and `_total_sponsored_contributors` stats because they're not reliable or meaningfullly different from the "pledged" stats. That effectively reverts most of c1ce72d4476464f77edce20edcf9b0f86bd91245.

The remaining variables were renamed for clarity.

See #198

This will be further clarified by #200, because it will change the labels that are displayed to viewers.