Email: Set `From` header to `bounce@` to reduce confusion
Sometimes we do want folks to reply to messages, and the `donotreply` address contradicts that. `bounce@` has a similar meaning, but is less known. `bounce@` has the advantage of matching the `Return-Path`, which may reduce the chance of a message being flagged as spam for forging the `From` header.

See https://github.com/WordPress/five-for-the-future/pull/206#discussion_r938418815